+++ src/main/java/org/opentripplanner/middleware/bugsnag/BugsnagJobs.java
@@ -9,7 +9,7 @@ import org.slf4j.LoggerFactory;
 
 import java.util.concurrent.TimeUnit;
 
-import static org.opentripplanner.middleware.spark.Main.getConfigPropertyAsInt;
+import static org.opentripplanner.middleware.OtpMiddlewareMain.getConfigPropertyAsInt;
 
 /**
  * Schedule the Bugsnag jobs. Only ten API requests can be made to Bugsnag within a minute. If additional jobs are added
